requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.32.3
beautifulsoup4==4.12.3
flask-limiter==3.5.0
validators==0.22.0
lxml==5.2.2
